Consolidate friendly progress card internals
